+++ src/control/mod.rs
@@ -1,0 +1,10 @@
+// espera::control
+//
+//! Time control.
+//
+
+mod looper;
+mod rate;
+mod sleeper;
+
+pub use {looper::*, rate::*, sleeper::*};